requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic
python-dotenv
numpy
pandas
scikit-learn
torch
psycopg2-binary
ultralytics
torchvision
PIL
sklearn